# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Localhost"
    },
    {
      "url": "http://host.docker.internal:3000/",
      "description": "docker host"
    }
  ],
  "paths": {
    "/api/users/": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userSchema"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/userSchema"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userSchemaInput"
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "Array of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userSchema"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userSchema"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/userSchema"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userSchema"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/userSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userSchemaInput"
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/data/bikelanes": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "GeoJSON Data",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/data/search": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "userSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 0
          },
          "username": {
            "type": "string",
            "example": "Neo Felipez"
          },
          "mail": {
            "type": "string",
            "example": "matrix@test.net"
          }
        },
        "required": [
          "id",
          "username",
          "mail"
        ],
        "xml": {
          "name": "userSchema"
        }
      },
      "userSchemaInput": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "Neo Felipez"
          },
          "mail": {
            "type": "string",
            "example": "matrix@test.net"
          }
        },
        "required": [
          "username",
          "mail"
        ],
        "xml": {
          "name": "userSchemaInput"
        }
      }
    }
  }
}